test(favorit): move delete favorit tests to UserServiceTest
